Completed basic 5 turn rock paper scissors JS game
Moved the choice functions inside the playRound functions to endable
the choices to be called each round instead of 1 time only.  Changed
the console log to include definitions for user. Completed basic
game, as the instructions said to resist the urge to update, tinker,
and attempt to perfect the game. It is at it's most basic functioning
state at this time.